requirements.txt
amqp==5.1.1
asgi-redis==1.1.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
dj-database-url==2.1.0
Django==4.2.6
django-birthday==0.1.4
django-cors-headers==4.3.0
django-filter==23.3
django-location-field==2.7.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
msgpack==1.0.7
msgpack-python==0.5.6
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.8
zope.interface==6.1
